+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Make span not to break the formatting
 * Add "Later" section for tasks that are not due today
 * Helper classes to task blocks for Day Planner for coloring blocks
+* Sync Google Calendar to Todoist
 
 ### 1.0.0: 2025-01-12